Conflicts:
	arch/arm/common/source/ArchMemory.cpp
	arch/arm/common/source/usb/platform.c
	arch/x86/64/include/ArchThreads.h
	arch/x86/64/source/ArchThreads.cpp
	common/include/ustl/cmemlink.h
	common/include/ustl/ustring.h

+++ common/include/fs/FileDescriptor.h
@@ -18,7 +18,7 @@ class FileDescriptor
     /**
      * the file descriptor
      */
-    uint32 fd_;
+    size_t fd_;
 
     /**
      * the file object
